+++ src/ai_risk_manager/collectors/plugins/__init__.py
@@ -5,6 +5,7 @@ from ai_risk_manager.collectors.plugins.base import (
     StackId,
     StackProbeResult,
 )
+from ai_risk_manager.collectors.plugins.django import DjangoCollectorPlugin
 from ai_risk_manager.collectors.plugins.fastapi import FastAPICollectorPlugin
 from ai_risk_manager.collectors.plugins.registry import (
     get_default_plugin,
@@ -17,6 +18,7 @@ __all__ = [
     "ArtifactBundle",
     "CollectorPlugin",
     "DetectionConfidence",
+    "DjangoCollectorPlugin",
     "FastAPICollectorPlugin",
     "StackId",
     "StackProbeResult",